[Home] fixes loading order
