+++ src/plugin/home/Resources/modules/tools/home/components/tool.jsx
@@ -2,11 +2,12 @@ import React from 'react'
 import {PropTypes as T} from 'prop-types'
 import get from 'lodash/get'
 
-import {Tool} from '#/main/core/tool'
+import {Tool, ToolPage} from '#/main/core/tool'
 
-import {PlayerMain} from '#/plugin/home/tools/home/player/containers/main'
 import {HomeEditor} from '#/plugin/home/tools/home/editor/containers/main'
-import {getTabSummary} from '#/plugin/home/tools/home/utils'
+import {flattenTabs, getTabSummary} from '#/plugin/home/tools/home/utils'
+import {HomeTab} from '#/plugin/home/tools/home/containers/tab'
+import {Tab as TabTypes} from '#/plugin/home/prop-types'
 
 const HomeTool = props => {
   const tabs = props.tabs
@@ -19,15 +20,42 @@ const HomeTool = props => {
       styles={['claroline-distribution-plugin-home-home-tool']}
       menu={1 < tabs.length ? tabs : []}
       editor={HomeEditor}
+      redirect={[
+        props.tabs[0] && {from: '/', exact: true, to: '/' + props.tabs[0].slug}
+      ].filter(redirect => !!redirect)}
+      pages={[
+        {
+          path: '/:slug',
+          onEnter: (params = {}) => props.setCurrentTab(params.slug),
+          render: (routeProps) => {
+            const flattened = flattenTabs(props.tabs)
+            if (flattened.find(tab => tab.slug === routeProps.match.params.slug)) {
+              return <HomeTab />
+            }
+
+            // tab does not exist
+            // let redirection open the first available
+            routeProps.history.replace(props.path)
+
+            return null
+          }
+        }
+      ]}
     >
-      <PlayerMain/>
+      {!props.loaded &&
+        <ToolPage />
+      }
     </Tool>
   )
 }
 
 HomeTool.propTypes = {
   path: T.string.isRequired,
-  tabs: T.array
+  loaded: T.bool.isRequired,
+  tabs: T.arrayOf(T.shape(
+    TabTypes.propTypes
+  )),
+  setCurrentTab: T.func.isRequired
 }
 
 HomeTool.defaultProps = {
