--- src/plugin/home/Resources/modules/tools/home/player/store/index.js
@@ -1,10 +1,0 @@
-
-import {actions} from '#/plugin/home/tools/home/player/store/actions'
-import {reducer} from '#/plugin/home/tools/home/player/store/reducer'
-import {selectors} from '#/plugin/home/tools/home/player/store/selectors'
-
-export {
-  actions,
-  reducer,
-  selectors
-}
